@@ -9,8 +9,8 @@
     "esp-js": "^6.5.1",
     "esp-js-react": "^6.5.1",
     "immutable": "^4.0.0-rc.12",
-    "react": "16.14.0",
-    "react-dom": "16.14.0"
+    "react": "18.3.1",
+    "react-dom": "18.3.1"
   },
   "devDependencies": {
     "jest": "26.0.1",
