@@ -10,8 +10,8 @@
     "esp-js-polimer": "^6.5.1",
     "esp-js-react": "^6.5.1",
     "esp-js-ui": "^6.5.1",
-    "react": "16.14.0",
-    "react-dom": "16.14.0",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
     "rxjs": "6.6.6",
     "uuid": "8.3.2"
   },
