@@ -27,7 +27,7 @@
     "esp-js": "^6.5.1",
     "esp-js-polimer": "^6.5.1",
     "prop-types": "^15.7.2",
-    "react": "16.14.0"
+    "react": "18.3.1"
   },
   "devDependencies": {
     "jest": "26.0.1",
