@@ -31,7 +31,7 @@
     "esp-js-react": "^6.5.1",
     "esp-js-rx": "^6.5.1",
     "query-string": "6.13.1",
-    "react": "16.14.0",
+    "react": "18.3.1",
     "rxjs": "6.6.6",
     "uuid": "8.3.2"
   },
