@@ -1,10 +1,10 @@
 import * as React from 'react';
-import * as ReactDOM from 'react-dom';
+import {createRoot} from 'react-dom/client';
 import {AppShell} from './shell/appShell';
 
 let appShell = new AppShell();
 appShell.start();
-ReactDOM.render(
-    appShell.rootElement,
-    document.getElementById('root')
-);
+
+const container = document.getElementById('root');
+const root = createRoot(container!);
+root.render(appShell.rootElement);
